+++ tests/grav/index.php
@@ -1,0 +1,1 @@
+<?php echo "grav front ", $_SERVER["REQUEST_URI"];