+++ tests/grav/system/defines.php
@@ -1,0 +1,1 @@
+<?php define("GRAV_VERSION", "1.7.48");